requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.3
python-dotenv==1.0.0
pymongo==4.6.3
bcrypt==3.2.0
PyJWT==2.8.0p